Refactoring: Everything to float, round to 1 decimal place, change deploy script - images

+++ database/migrations/2021_12_12_133442_create_product_warehouse_table.php
@@ -16,8 +16,8 @@ class CreateProductWarehouseTable extends Migration
         Schema::create('product_warehouse', function (Blueprint $table) {
             $table->primary(['product_id', 'warehouse_id']);
 
-            $table->float('amount');
-            $table->decimal('price');
+            $table->float('amount', 8, 1);
+            $table->float('price', 8, 1);
 
             $table->foreignId('product_id')->constrained('products');
             $table->foreignId('warehouse_id')->constrained('warehouses');
